requirements.txt
ffmpeg-python 
opencv-python